This was taking a ludicrous amount of time in Dagger's test suite (more generally on every session start, but Dagger's test suite is sort of a stress test for that).

Simply appending seems to work fine, even with duplicate type definitions. Considering the performance improvement, if we do end up running into any trouble it seems worth trying an alternative fix on top of this change rather than reverting it.

Performance comparison (GQL optimization = this change, ID refactor = https://github.com/dagger/dagger/pull/4973):

- no ID refactor, no GQL optimization
    - no profiling
        - DONE 359 tests in 211.938s
        - DONE 359 tests in 192.635s
    - profiling
        - DONE 359 tests in 203.365s
- yes ID refactor, no GQL optimization
    - DONE 359 tests, 1 skipped in 187.724s
- yes ID refactor, yes GQL optimization
    - profiling
        - DONE 359 tests, 1 skipped in 63.278s
        - DONE 359 tests, 1 skipped in 67.219s
        - DONE 359 tests, 1 skipped in 68.321s
        - DONE 359 tests, 1 skipped in 63.550s
    - no profiling
        - DONE 359 tests, 1 skipped in 60.048s
        - DONE 359 tests, 1 skipped in 59.450s
- no ID refactor, yes GQL optimization
    - DONE 359 tests in 97.016s
    - DONE 359 tests in 69.192s
    - DONE 359 tests in 88.365s
    - DONE 359 tests in 66.448s

Flamegraph before + after:

![image](https://user-images.githubusercontent.com/1880/232898234-6ec72cf2-9a82-4f7d-84a7-1dad3b50409a.png)

![image](https://user-images.githubusercontent.com/1880/232898270-e05450ff-b8a2-4209-8247-95f9c467afb2.png)